data_producer/requirements.txt
kafka-python
docopt